Add description to Accessbility Service
