+++ app/src/main/java/com/example/android/minutelauncher/MinuteAccessibilityService.kt
@@ -1,8 +1,12 @@
 package com.example.android.minutelauncher
 
 import android.accessibilityservice.AccessibilityService
+import android.accessibilityservice.AccessibilityServiceInfo
+import android.content.Context
 import android.content.Intent
+import android.content.pm.ServiceInfo
 import android.view.accessibility.AccessibilityEvent
+import android.view.accessibility.AccessibilityManager
 
 class MinuteAccessibilityService: AccessibilityService() {
 
@@ -29,4 +33,21 @@ class MinuteAccessibilityService: AccessibilityService() {
     a = null
     return super.onUnbind(intent)
   }
+}
+
+fun isAccessibilityServiceEnabled(
+  context: Context,
+  service: Class<out AccessibilityService?>
+): Boolean {
+  val am = context.getSystemService(Context.ACCESSIBILITY_SERVICE) as (AccessibilityManager)
+  val enabledServices: List<AccessibilityServiceInfo> =
+    am.getEnabledAccessibilityServiceList(AccessibilityServiceInfo.FEEDBACK_ALL_MASK)
+  for (enabledService in enabledServices) {
+    val enabledServiceInfo: ServiceInfo = enabledService.resolveInfo.serviceInfo
+    if (enabledServiceInfo.packageName.equals(context.packageName) && enabledServiceInfo.name.equals(
+        service.name
+      )
+    ) return true
+  }
+  return false
 }
