+++ app/src/main/res/xml/minute_accessibility_service_config.xml
@@ -1,5 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <accessibility-service xmlns:android="http://schemas.android.com/apk/res/android"
+    android:description="@string/accessibility_service_description"
     android:packageNames="com.android.settings"
     android:accessibilityEventTypes="typeWindowStateChanged"
     android:accessibilityFlags="flagDefault"
